Move benchmark code into a separate source
